requirements.txt
Django==2.0.5
djangorestframework==3.8.2
Pillow==5.1.0
pylint==1.8.4
pylint-django==0.11
Markdown==2.6.11
